Added the GameSettings.ini file.
If the "MathsChess" folder in Documents is missing, it'll be automatically created, same with the "GameSettings.ini file". On the first time setup, it'll randomly generate a "Guest" name for you, you can change this in settings. Also will sync your username with the server. To-do same from the last to-do list.